:truck: [#4606] Move CatalogueSerializer to shared library code
Objects API and ZGW APIs registration backends both use the same
option structure for the catalogue, so we can re-use the serializer to
have a single source of truth.
